LayoutTests/media/media-source/media-source-real-rapid-seeks.html

<!DOCTYPE html>
<html>
<head>
    <title>media-source-real-rapid-seeks</title>
    <script src="mp4-generator.js"></script>
    <script src="../video-test.js"></script>
    <script>
    // Tests rapid sequential seeks after endOfStream. This exercises the
    // cross-process seek state machine — a pattern that exposed timing
    // bugs where currentTime briefly went backwards or reenqueue flags
    // got stuck during the re-architecture.

    var source;
    var sourceBuffer;

    async function runTest() {
        findMediaElement();

        source = new MediaSource();
        video.src = URL.createObjectURL(source);
        await waitFor(source, 'sourceopen');

        sourceBuffer = source.addSourceBuffer(MP4.VIDEO_TYPE);

        var initData = MP4.initSegment({
            timescale: 1000,
            tracks: [{ id: 1, type: 'video', timescale: 1000 }]
        });
        sourceBuffer.appendBuffer(initData);
        await waitFor(sourceBuffer, 'updateend');

        // 8 seconds of video, sync every 2 seconds (at 0, 2, 4, 6)
        var media = MP4.mediaSegment({
            sequenceNumber: 1,
            tracks: [{ id: 1, type: 'video', baseDecodeTime: 0, samples: [
                { duration: 1000, compositionTimeOffset: 0, isSync: true },
                { duration: 1000, compositionTimeOffset: 0, isSync: false },
                { duration: 1000, compositionTimeOffset: 0, isSync: true },
                { duration: 1000, compositionTimeOffset: 0, isSync: false },
                { duration: 1000, compositionTimeOffset: 0, isSync: true },
                { duration: 1000, compositionTimeOffset: 0, isSync: false },
                { duration: 1000, compositionTimeOffset: 0, isSync: true },
                { duration: 1000, compositionTimeOffset: 0, isSync: false },
            ]}]
        });
        sourceBuffer.appendBuffer(media);
        await waitFor(sourceBuffer, 'updateend');

        source.endOfStream();

        // Rapid seeks: forward, backward, forward
        run('video.currentTime = 2');
        await waitFor(video, 'seeked');
        testExpectedEqualWithTolerance('video.currentTime', 2, 0.1);

        run('video.currentTime = 6');
        await waitFor(video, 'seeked');
        testExpectedEqualWithTolerance('video.currentTime', 6, 0.1);

        run('video.currentTime = 0');
        await waitFor(video, 'seeked');
        testExpectedEqualWithTolerance('video.currentTime', 0, 0.1);

        run('video.currentTime = 4');
        await waitFor(video, 'seeked');
        testExpectedEqualWithTolerance('video.currentTime', 4, 0.1);

        endTest();
    }
    </script>
</head>
<body onload="runTest()">
    <video></video>
</body>
</html>
